pyproject.toml
[tool.poetry]
name = "anitrend-relations-py"
version = "0.1.0"
description = ""
authors = ["Maxwell <mxt.developer@gmail.com>"]
license = "GPL"
readme = "README.md"
packages = [{include = "anitrend_relations_py"}]

[tool.poetry.dependencies]
python = "^3.10"
uplink = "0.9.7"
requests = "2.31.0"
marshmallow = "3.19.0"
python-decouple = "3.8"
django-filter = "23.2"
graphene-django = "3.1.1"
graphene-django-extras = { git = "https://github.com/wax911/graphene-django-extras.git", tag = "0.1.0" }
django-cors-headers = "4.0.0"
graphene = "3.2.2"
django-enumfield = "3.1"
django-graphql-playground = "1.0"
django-extensions = "3.2.1"
django-q = "1.3.9"
django = "4.2.1"
dependency_injector = "4.41.0"
serde = "0.9.0"
gunicorn = "20.1.0"
psycopg2-binary = "2.9.6"
dj_database_url = "1.2.0"
reactivex = "4.0.4"
pyxtension = "1.16.1"
mongoengine = "0.27.0"
graphene-mongo = "0.2.15"

[tool.poetry.group.test.dependencies]
mongomock = "3.16.0"
pytest = "7.3.1"
pytest-django = "4.5.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
